+++ lib/Adaptor.js
@@ -4,7 +4,6 @@ Object.defineProperty(exports, "__esModule", {
   value: true
 });
 exports.execute = execute;
-exports.nestArray = nestArray;
 exports.create = create;
 exports.update = update;
 exports.getTEIs = getTEIs;
@@ -203,85 +202,6 @@ _axios.default.interceptors.response.use(function (response) {
 
   return Promise.reject(error);
 });
-
-function expandAndSetOperation(options, state, operationName) {
-  return {
-    operationName,
-    ...(0, _languageCommon.expandReferences)(options)(state)
-  };
-}
-
-const isArray = variable => !!variable && variable.constructor === Array;
-
-function nestArray(data, key) {
-  return isArray(data) ? {
-    [key]: data
-  } : data;
-}
-
-function log(operationName, apiVersion, url, resourceType, params) {
-  (0, _Utils.logOperation)(operationName);
-  (0, _Utils.logApiVersion)(apiVersion);
-  (0, _Utils.logWaitingForServer)(url, params);
-  (0, _Utils.warnExpectLargeResult)(resourceType, url);
-}
-
-function extractValuesForAxios(operationName, values) {
-  return state => {
-    var _values$options$apiVe, _values$options, _values$options2;
-
-    const apiVersion = (_values$options$apiVe = (_values$options = values.options) === null || _values$options === void 0 ? void 0 : _values$options.apiVersion) !== null && _values$options$apiVe !== void 0 ? _values$options$apiVe : state.configuration.apiVersion;
-    const {
-      username,
-      password,
-      hostUrl
-    } = state.configuration;
-    const auth = {
-      username,
-      password
-    };
-    let urlString = '/' + values.resourceType;
-
-    if (operationName === 'update') {
-      urlString += '/' + values.path;
-    }
-
-    const url = (0, _Utils.buildUrl)(urlString, hostUrl, apiVersion);
-    const urlParams = new URLSearchParams((_values$options2 = values.options) === null || _values$options2 === void 0 ? void 0 : _values$options2.params);
-    return {
-      resourceType: values.resourceType,
-      data: values.data,
-      apiVersion,
-      auth,
-      url,
-      urlParams,
-      callback: values.callback
-    };
-  };
-}
-
-function expandExtractAndLog(operationName, initialParams) {
-  return state => {
-    const {
-      resourceType,
-      data,
-      apiVersion,
-      auth,
-      url,
-      urlParams,
-      callback
-    } = extractValuesForAxios(operationName, (0, _languageCommon.expandReferences)(initialParams)(state))(state);
-    log(operationName, apiVersion, url, resourceType, urlParams);
-    return {
-      url,
-      data,
-      resourceType,
-      auth,
-      urlParams,
-      callback
-    };
-  };
-}
 /**
  * Create a record
  * @public
@@ -410,8 +330,8 @@ function create(resourceType, data, options, callback) {
       auth,
       urlParams,
       callback
-    } = expandExtractAndLog('create', initialParams)(state);
-    return (0, _Client.post)(url, nestArray(data, resourceType), {
+    } = (0, _Utils.expandExtractAndLog)('create', initialParams)(state);
+    return (0, _Client.post)(url, (0, _Utils.nestArray)(data, resourceType), {
       auth,
       params: urlParams
     }).then(result => {
@@ -585,7 +505,7 @@ function update(resourceType, path, data, options, callback) {
       auth,
       _,
       callback
-    } = expandExtractAndLog('update', initialParams)(state);
+    } = (0, _Utils.expandExtractAndLog)('update', initialParams)(state);
     return (0, _Client.put)(url, data, {
       auth
     }).then(result => {
@@ -618,7 +538,7 @@ function update(resourceType, path, data, options, callback) {
 
 function getTEIs(params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'getTEIs');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'getTEIs');
     return getData('trackedEntityInstances', params, expandedOptions, callback)(state);
   };
 }
@@ -662,7 +582,7 @@ function upsertTEI(uniqueAttributeId, data, options, callback) {
 
     uniqueAttributeId = (0, _languageCommon.expandReferences)(uniqueAttributeId)(state);
     const body = (0, _languageCommon.expandReferences)(data)(state);
-    const expandedOptions = expandAndSetOperation(options, state, 'upsertTEI');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'upsertTEI');
     const {
       password,
       username,
@@ -752,7 +672,7 @@ function upsertTEI(uniqueAttributeId, data, options, callback) {
 
 function getEvents(params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'getEvents');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'getEvents');
     return getData('events', params, expandedOptions, callback)(state);
   };
 }
@@ -771,7 +691,7 @@ function getEvents(params, options, callback) {
 
 function getPrograms(params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'getPrograms');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'getPrograms');
     return getData('programs', params, expandedOptions, callback)(state);
   };
 }
@@ -795,7 +715,7 @@ function getPrograms(params, options, callback) {
 
 function getEnrollments(params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'getEnrollments');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'getEnrollments');
     return getData('enrollments', params, expandedOptions, callback)(state);
   };
 }
@@ -821,7 +741,7 @@ function getEnrollments(params, options, callback) {
 
 function updateEnrollments(path, data, params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'updateEnrollments');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'updateEnrollments');
     return update('enrollments', path, data, params, expandedOptions, callback)(state);
   };
 }
@@ -844,7 +764,7 @@ function cancelEnrollment(enrollmentId, params, options, callback) {
   return state => {
     enrollmentId = (0, _languageCommon.expandReferences)(enrollmentId)(state);
     const path = `${enrollmentId}/cancelled`;
-    const expandedOptions = expandAndSetOperation(options, state, 'cancelEnrollment');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'cancelEnrollment');
     return update('enrollments', path, null, params, expandedOptions, callback)(state);
   };
 }
@@ -867,7 +787,7 @@ function completeEnrollment(enrollmentId, params, options, callback) {
   return state => {
     enrollmentId = (0, _languageCommon.expandReferences)(enrollmentId)(state);
     const path = `${enrollmentId}/completed`;
-    const expandedOptions = expandAndSetOperation(options, state, 'enrollments');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'enrollments');
     return update('enrollments', path, null, params, expandedOptions, callback)(state);
   };
 }
@@ -888,7 +808,7 @@ function completeEnrollment(enrollmentId, params, options, callback) {
 
 function getRelationships(params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'getRelationships');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'getRelationships');
     return getData('relationships', params, expandedOptions, callback)(state);
   };
 }
@@ -914,7 +834,7 @@ function getRelationships(params, options, callback) {
 
 function getDataValues(params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'getDataValues');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'getDataValues');
     return getData('dataValueSets', params, expandedOptions, callback)(state);
   };
 }
@@ -935,7 +855,7 @@ function generateDhis2UID(options, callback) {
   return state => {
     var _options$limit;
 
-    const expandedOptions = expandAndSetOperation(options, state, 'generateDhis2UID');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'generateDhis2UID');
     const limit = {
       limit: (_options$limit = options === null || options === void 0 ? void 0 : options.limit) !== null && _options$limit !== void 0 ? _options$limit : 1
     };
@@ -1027,7 +947,7 @@ function discover(httpMethod, endpoint) {
 
 function getAnalytics(params, options, callback) {
   return state => {
-    const expandedOptions = expandAndSetOperation(options, state, 'getAnalytics');
+    const expandedOptions = (0, _Utils.expandAndSetOperation)(options, state, 'getAnalytics');
     return getData(`analytics`, params, expandedOptions, callback)(state);
   };
 }
