+++ lib/Client.js
@@ -11,9 +11,6 @@ var _axios = _interopRequireDefault(require("axios"));
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function post(url, data, configs) {
-  console.log('URL', url);
-  console.log('DATA', data);
-  console.log('CONFIGS', configs);
   return _axios.default.request({
     method: 'POST',
     url,
