+++ lib/Utils.js
@@ -18,11 +18,16 @@ exports.getIndicesOf = getIndicesOf;
 exports.isLike = isLike;
 exports.applyFilter = applyFilter;
 exports.parseFilter = parseFilter;
+exports.expandAndSetOperation = expandAndSetOperation;
+exports.nestArray = nestArray;
+exports.expandExtractAndLog = expandExtractAndLog;
 exports.CONTENT_TYPES = exports.dhis2OperatorMap = exports.Log = void 0;
 
 var _lodash = require("lodash");
 
 var _axios = _interopRequireDefault(require("axios"));
+
+var _languageCommon = require("@openfn/language-common");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -181,6 +186,85 @@ function parseFilter(filterExpression) {
   const filterTokens = filterExpression === null || filterExpression === void 0 ? void 0 : filterExpression.split(':');
   filterTokens ? filterTokens[1] = dhis2OperatorMap[(_filterTokens$ = filterTokens[1]) !== null && _filterTokens$ !== void 0 ? _filterTokens$ : null] : null;
   return filterTokens;
+}
+
+function expandAndSetOperation(options, state, operationName) {
+  return {
+    operationName,
+    ...(0, _languageCommon.expandReferences)(options)(state)
+  };
+}
+
+const isArray = variable => !!variable && variable.constructor === Array;
+
+function nestArray(data, key) {
+  return isArray(data) ? {
+    [key]: data
+  } : data;
+}
+
+function log(operationName, apiVersion, url, resourceType, params) {
+  logOperation(operationName);
+  logApiVersion(apiVersion);
+  logWaitingForServer(url, params);
+  warnExpectLargeResult(resourceType, url);
+}
+
+function extractValuesForAxios(operationName, values) {
+  return state => {
+    var _values$options$apiVe, _values$options, _values$options2;
+
+    const apiVersion = (_values$options$apiVe = (_values$options = values.options) === null || _values$options === void 0 ? void 0 : _values$options.apiVersion) !== null && _values$options$apiVe !== void 0 ? _values$options$apiVe : state.configuration.apiVersion;
+    const {
+      username,
+      password,
+      hostUrl
+    } = state.configuration;
+    const auth = {
+      username,
+      password
+    };
+    let urlString = '/' + values.resourceType;
+
+    if (operationName === 'update') {
+      urlString += '/' + values.path;
+    }
+
+    const url = buildUrl(urlString, hostUrl, apiVersion);
+    const urlParams = new URLSearchParams((_values$options2 = values.options) === null || _values$options2 === void 0 ? void 0 : _values$options2.params);
+    return {
+      resourceType: values.resourceType,
+      data: values.data,
+      apiVersion,
+      auth,
+      url,
+      urlParams,
+      callback: values.callback
+    };
+  };
+}
+
+function expandExtractAndLog(operationName, initialParams) {
+  return state => {
+    const {
+      resourceType,
+      data,
+      apiVersion,
+      auth,
+      url,
+      urlParams,
+      callback
+    } = extractValuesForAxios(operationName, (0, _languageCommon.expandReferences)(initialParams)(state))(state);
+    log(operationName, apiVersion, url, resourceType, urlParams);
+    return {
+      url,
+      data,
+      resourceType,
+      auth,
+      urlParams,
+      callback
+    };
+  };
 }
 
 const CONTENT_TYPES = {
